refactor: rename epoch_storage.rs
