+++ crates/hashi/src/storage.rs
